+++ schema/Url.py
@@ -1,0 +1,5 @@
+from marshmallow import fields
+from database.db import ma
+
+class UrlSchema(ma.Schema):
+    url = fields.Url(required=True)
