+++ server.py
@@ -1,0 +1,7 @@
+# Initialize flask
+
+from flask import Flask
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database/app.db'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
